frontend/requirements.txt
Flask>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
Werkzeug>=3.0.0
python-dotenv>=1.0.0
